dev-requirements.txt
mypy
django-stubs
mypy-django-plugin
pytest
pytest-cov
black
isort
types-requests
# Added from main requirements.txt previously
django-pytest 
pytest-django
coverage